requirements.txt
pyyaml
django-dotenv
boto3
markdown